Characters now save info better to DB and retrieve Character portraits now stream from internet Added Delete Account button that does NOTHING!

+++ EveApi/Api/Interfaces/ICharacterInfo.cs
@@ -75,7 +75,7 @@ namespace EveApi.Api.Interfaces
         /// <summary>
         /// Amount of ISK the character has. 
         /// </summary>
-        decimal AccountBalance
+        double AccountBalance
         {
             get;
             set;
